requirements.txt
playwright
pandas
beautifulsoup4
lxml
google-cloud-storage